chore: Bumped wgpu to 22.0
This required bumping the git submodules. The git submodules had to be bumped as well.
With this bump, the imgui-winit-support crate was moved out of imgui-rs upstream,
so I added it as a separate git submodule. I also forked imgui-wgpu to bump it to
wgpu 22.0, updating this git submodule to point to my fork (hopefully changes will
go upstream soon). I've put dependencies into the workspace.dependencies table
in Cargo.toml, as there were several issues with multiple versions being around.